refactoring time deltas into utils module
